requirements.txt
flet
sqlalchemy
jinja2
python-dotenv
python-ulid
openai
google-generativeai
